demo-chatbot/requirements.txt
flask
twilio
python-dotenv
openai
pyngrok
faster-whisper
numpy